updated environment, fix float issue

+++ chemberta/utils/roberta_regression.py
@@ -119,7 +119,9 @@ class RobertaForSequenceClassification(RobertaPreTrainedModel):
                 loss = loss_fct(logits.view(-1), labels.view(-1))
             else:
                 loss_fct = CrossEntropyLoss()
-                loss = loss_fct(logits.view(-1, self.num_labels), labels.view(-1))
+                loss = loss_fct(
+                    logits.view(-1, self.num_labels), labels.long().view(-1)
+                )
 
         if not return_dict:
             output = (logits,) + outputs[2:]
